requirements.txt
Flask
Flask-SocketIO
paho-mqtt
PyJWT
Werkzeug